requirements.txt
python-dotenv==1.2.1
tqdm==4.65.0
networkx==3.1
matplotlib.pyplot==3.8.0
scipy==1.16.3